refactor(core): isolate SCXML runtime
