+++ .changeset/strict-scxml-conformance.md
@@ -2,4 +2,10 @@
 'xstate': patch
 ---
 
-Improve strict SCXML conformance for converted machines, including transition selection, executable content, datamodel evaluation, invocation, event metadata, and completion behavior.
+Add `createMachineFromSCXML(...)` through the `xstate/scxml` entry point. Converted machines follow strict SCXML behavior for transition selection, executable content, datamodel evaluation, invocation, event metadata, and completion.
+
+```ts
+import { createMachineFromSCXML } from 'xstate/scxml';
+
+const machine = createMachineFromSCXML(scxml);
+```
