+++ packages/core/scxml/package.json
@@ -1,0 +1,8 @@
+{
+  "main": "../dist/xstate-scxml.cjs.js",
+  "module": "../dist/xstate-scxml.esm.js",
+  "umd:main": "../dist/xstate-scxml.umd.min.js",
+  "preconstruct": {
+    "umdName": "XStateSCXML"
+  }
+}
