+++ packages/core/src/createMachineFromConfig.ts
@@ -1,14 +1,12 @@
 import {
   Action,
   AnyActorLogic,
-  AnyActorRef,
   AnyEventObject,
   AnyStateMachine,
   EventObject,
   MachineContext,
   MetaObject
 } from './types';
-import { Next_StateNodeConfig } from './types.v6';
 import { createMachine } from './createMachine';
 import { parseDelayToMilliseconds } from './delay';
 
@@ -18,48 +16,14 @@ function delayToMs(delay: string | number): number {
   return typeof delay === 'string' ? parseFloat(delay) || 0 : delay;
 }
 
-function normalizeScxmlEventType(eventType: string): string {
-  return /^error(\.|$)/.test(eventType) ? `xstate.${eventType}` : eventType;
-}
-
-function toScxmlEventName(eventType: string): string {
-  return eventType.startsWith('xstate.error.')
-    ? eventType.slice('xstate.'.length)
-    : eventType;
-}
-
-function matchesScxmlEventDescriptor(
-  descriptor: string,
-  event: EventObject
-): boolean {
-  if (descriptor === '*') return true;
-  if (descriptor.endsWith('.*')) {
-    const prefix = descriptor.slice(0, -2);
-    return event.type === prefix || event.type.startsWith(`${prefix}.`);
-  }
-  if (descriptor.startsWith('xstate.done.actor.')) {
-    return (
-      event.type === 'xstate.done.actor' &&
-      (event as any).actorId === descriptor.slice('xstate.done.actor.'.length)
-    );
-  }
-  if (descriptor.startsWith('xstate.done.state.')) {
-    return (
-      event.type === 'xstate.done.state' &&
-      (event as any).stateId === descriptor.slice('xstate.done.state.'.length)
-    );
-  }
-  return event.type === descriptor || event.type.startsWith(`${descriptor}.`);
-}
-
-export interface RaiseJSON {
+interface RaiseJSON {
   type: '@xstate.raise';
   event: EventObject;
   id?: string;
   delay?: number;
 }
 
-export interface CancelJSON {
+interface CancelJSON {
   type: '@xstate.cancel';
   id: string;
 }
@@ -67,12 +31,6 @@ export interface CancelJSON {
 interface LogJSON {
   type: '@xstate.log';
   args: any[];
-}
-
-export interface ScxmlLogJSON {
-  type: 'scxml.log';
-  label?: string;
-  expr?: string;
 }
 
 interface EmitJSON {
@@ -83,116 +41,6 @@ interface EmitJSON {
 interface AssignJSON {
   type: '@xstate.assign';
   context: MachineContext;
-}
-
-interface ScxmlAssignJSON {
-  type: 'scxml.assign';
-  /** SCXML location attribute - the context property to assign to */
-  location: string;
-  /** SCXML expr attribute - expression to evaluate */
-  expr: string;
-}
-
-export interface ScxmlRaiseJSON {
-  type: 'scxml.raise';
-  /** Event type, or undefined if using eventexpr */
-  event?: string;
-  /** Expression to evaluate for event type */
-  eventexpr?: string;
-  /** Params with expressions to evaluate */
-  params?: Array<{ name: string; expr?: string; location?: string }>;
-  namelist?: string[];
-  id?: string;
-  idlocation?: string;
-  delay?: number;
-  /** Expression for delay */
-  delayexpr?: string;
-  /** Static target (e.g. '#_parent') */
-  target?: string;
-  /** Expression for target */
-  targetexpr?: string;
-  processorType?: string;
-  content?: ScxmlContentJSON;
-}
-
-export interface ScxmlXmlJSON {
-  name: string;
-  attributes?: Record<string, string>;
-  children?: ScxmlXmlJSON[];
-}
-
-export interface ScxmlContentJSON {
-  expr?: string;
-  text?: string;
-  xml?: ScxmlXmlJSON;
-}
-
-interface ScxmlScriptJSON {
-  type: 'scxml.script';
-  /** The script code to execute */
-  code: string;
-  global?: boolean;
-}
-
-export interface ScxmlForeachJSON {
-  type: 'scxml.foreach';
-  array: string;
-  item: string;
-  index?: string;
-  actions: ActionJSON[];
-}
-
-export interface ScxmlCancelJSON {
-  type: 'scxml.cancel';
-  sendidexpr: string;
-}
-
-export interface ScxmlDonedataJSON {
-  params?: Array<{ name: string; expr: string }>;
-  contentExpr?: string;
-  contentText?: string;
-}
-
-export interface ScxmlDataJSON {
-  id: string;
-  expr?: string;
-  content?: string;
-  src?: string;
-  xml?: ScxmlXmlJSON;
-}
-
-/**
- * Isolated executable-content block. Errors in nested actions stop this block
- * but do not propagate to the surrounding action list. Used to model SCXML's
- * separate <onentry>/<onexit> blocks: each is its own block per spec.
- */
-interface ScxmlBlockJSON {
-  type: 'scxml.block';
-  actions: ActionJSON[];
-}
-
-interface ScxmlDatamodelJSON {
-  type: 'scxml.datamodel';
-  data: ScxmlDataJSON[];
-}
-
-interface ScxmlForwardJSON {
-  type: 'scxml.forward';
-  invokeId: string;
-}
-
-interface ScxmlFinalizeJSON {
-  type: 'scxml.finalize';
-  invokeId: string;
-  actions: ActionJSON[];
-}
-
-interface ScxmlIfJSON {
-  type: 'scxml.if';
-  branches: Array<{
-    cond?: string;
-    actions: ActionJSON[];
-  }>;
 }
 
 type BuiltInActionJSON =
@@ -209,22 +57,7 @@ interface CustomActionJSON {
 
 export type ActionJSON =
   | CustomActionJSON
-  | RaiseJSON
-  | CancelJSON
-  | LogJSON
-  | EmitJSON
-  | AssignJSON
-  | ScxmlAssignJSON
-  | ScxmlRaiseJSON
-  | ScxmlScriptJSON
-  | ScxmlIfJSON
-  | ScxmlForeachJSON
-  | ScxmlCancelJSON
-  | ScxmlLogJSON
-  | ScxmlBlockJSON
-  | ScxmlDatamodelJSON
-  | ScxmlForwardJSON
-  | ScxmlFinalizeJSON
+  | BuiltInActionJSON
   | ExpressionJSON
   | CodeJSON;
 
@@ -244,7 +77,6 @@ interface CodeJSON {
 }
 
 type ResolvableJSON = ExpressionJSON | CodeJSON;
-
 type ConditionJSON = GuardJSON | ResolvableJSON;
 
 interface ChoiceBranchJSON {
@@ -267,11 +99,6 @@ export interface InvokeJSON {
   onSnapshot?: TransitionConfigJSON | TransitionConfigJSON[];
   timeout?: number | string | ResolvableJSON;
   onTimeout?: TransitionConfigJSON | TransitionConfigJSON[];
-  /** @internal SCXML invocation input expressions. */
-  _scxmlInput?: {
-    namelist?: string[];
-    params?: Array<{ name: string; expr?: string; location?: string }>;
-  };
 }
 
 export interface TransitionJSON {
@@ -284,12 +111,6 @@ export interface TransitionJSON {
   reenter?: boolean;
   meta?: MetaObject;
   input?: unknown;
-  /** @internal Strict semantics for transitions produced by the SCXML converter. */
-  _scxml?: {
-    type?: 'internal' | 'external';
-    eventDescriptors?: string[];
-    finalize?: Array<{ invokeId: string; actions: ActionJSON[] }>;
-  };
 }
 
 type TransitionConfigJSON = TransitionJSON | ResolvableJSON;
@@ -305,11 +126,6 @@ export interface StateNodeJSON {
   after?: Record<string, TransitionConfigJSON | TransitionConfigJSON[]>;
   always?: TransitionConfigJSON | TransitionConfigJSON[];
   choice?: ChoiceBranchJSON[] | ResolvableJSON;
-  /**
-   * JSON route config. Unlike the authoring API (a function), the JSON layer
-   * allows an object form whose `guard` may be a named guard reference (string)
-   * resolved against the machine's `guards` sources.
-   */
   route?:
     | {
         description?: string;
@@ -332,22 +148,11 @@ export interface StateNodeJSON {
   target?: string | [string, ...string[]];
   output?: unknown;
   context?: Record<string, unknown>;
-  _scxmlDonedata?: ScxmlDonedataJSON;
-  /** @internal Initial transition emitted by the SCXML converter. */
-  _scxmlInitial?: {
-    targets: string[];
-    actions?: ActionJSON[];
-  };
-  /** @internal Executable content on an SCXML history default transition. */
-  _scxmlHistoryActions?: ActionJSON[];
-  /** @internal Datamodel declarations initialized on entry. */
-  _scxmlData?: ScxmlDataJSON[];
 }
+
 export interface MachineJSON extends StateNodeJSON {
   '@exprLang'?: string;
   version?: string;
-  /** @internal All declared SCXML datamodel identifiers. */
-  _scxmlDataIds?: string[];
   actions?: Record<string, ActionJSON | ActionJSON[]>;
   guards?: Record<string, { when: ConditionJSON }>;
   actors?: Record<string, unknown>;
@@ -413,10 +218,6 @@ function isResolvable(value: unknown): value is ResolvableJSON {
 
 function isBuiltInActionType(type: string): boolean {
   return type.startsWith('@xstate.');
-}
-
-function isScxmlActionType(type: string): boolean {
-  return type.startsWith('scxml.');
 }
 
 function toPath(parent: string, key: string | number): string {
@@ -648,7 +449,7 @@ function assertMachineJSON(
     }
     if (
       !resolvedSources.guards[condition.type] &&
-      !['scxml.cond', 'xstate.stateIn', 'xstate.not'].includes(condition.type)
+      !['xstate.stateIn', 'xstate.not'].includes(condition.type)
     ) {
       throw new Error(`Missing guard source "${condition.type}"`);
     }
@@ -667,7 +468,7 @@ function assertMachineJSON(
       throw new Error(`Invalid action at ${path}`);
     }
     assertResolvable((action as CustomActionJSON).params, `${path}.params`);
-    if (isBuiltInActionType(action.type) || isScxmlActionType(action.type)) {
+    if (isBuiltInActionType(action.type)) {
       assertResolvable(action, path);
       return;
     }
@@ -768,10 +569,7 @@ function assertMachineJSON(
       const invokes = Array.isArray(node.invoke) ? node.invoke : [node.invoke];
       invokes.forEach((invoke, index) => {
         const invokePath = `${path}.invoke${Array.isArray(node.invoke) ? `[${index}]` : ''}`;
-        const extInv = invoke as InvokeJSON & {
-          _nestedMachineJSON?: MachineJSON;
-        };
-        if (!extInv._nestedMachineJSON && !resolvedSources.actors[invoke.src]) {
+        if (!resolvedSources.actors[invoke.src]) {
           throw new Error(`Missing actor source "${invoke.src}"`);
         }
         assertResolvable(invoke.input, `${invokePath}.input`);
@@ -843,246 +641,6 @@ function assertMachineJSON(
   assertStateNode(json, '$');
 }
 
-const SCXML_SYSTEM_VARIABLES = new Set([
-  '_sessionid',
-  '_name',
-  '_ioprocessors',
-  '_event',
-  '_xstateScxmlInputKeys'
-]);
-const scxmlScriptEnvironments = new WeakMap<
-  AnyActorRef,
-  { declarations: string[] }
->();
-const scxmlMachineNames = new WeakMap<AnyActorRef, string>();
-const scxmlIoProcessors = new WeakMap<AnyActorRef, object>();
-const scxmlEvents = new WeakMap<object, object>();
-const scxmlEnteringStates = new WeakMap<AnyActorRef, string[]>();
-
-function getScxmlIoProcessors(self: AnyActorRef, sessionId: string): object {
-  let processors = scxmlIoProcessors.get(self);
-  if (!processors) {
-    const processor = { location: `#_scxml_${sessionId}` };
-    processors = {
-      scxml: processor,
-      'http://www.w3.org/TR/scxml/#SCXMLEventProcessor': processor
-    };
-    scxmlIoProcessors.set(self, processors);
-  }
-  return processors;
-}
-
-function getFunctionDeclarations(code: string): string[] {
-  const declarations: string[] = [];
-  const pattern = /\bfunction\s+[$_a-zA-Z][$_a-zA-Z0-9]*\s*\([^)]*\)\s*\{/g;
-  for (const match of code.matchAll(pattern)) {
-    let depth = 1;
-    let index = match.index! + match[0].length;
-    while (index < code.length && depth) {
-      if (code[index] === '{') depth++;
-      if (code[index] === '}') depth--;
-      index++;
-    }
-    declarations.push(code.slice(match.index, index));
-  }
-  return declarations;
-}
-
-/** Evaluates an SCXML expression with context variables available via `with`. */
-function evaluateExpr(
-  context: MachineContext,
-  expr: string,
-  event: AnyEventObject | null,
-  self?: AnyActorRef
-): unknown {
-  const normalizedExpr = expr.replace(/;+\s*$/, '');
-  const sessionId = (self as any)?.sessionId ?? 'session_scxml';
-  const machineName = self
-    ? (scxmlMachineNames.get(self) ?? '(machine)')
-    : '(machine)';
-  const dataContext = Object.fromEntries(
-    Object.entries(context).filter(([key]) => !SCXML_SYSTEM_VARIABLES.has(key))
-  );
-  const fnBody = `
-${self ? (scxmlScriptEnvironments.get(self)?.declarations.join('\n') ?? '') : ''}
-with (context) {
-  return (${normalizedExpr});
-}
-  `.trim();
-  // eslint-disable-next-line @typescript-eslint/no-implied-eval
-  const fn = new Function(
-    'context',
-    '_event',
-    '_sessionid',
-    '_name',
-    '_ioprocessors',
-    'In',
-    fnBody
-  );
-  const In = (stateId: string) => {
-    if (!self) return false;
-    const sanitized = stateId.replace(/\./g, '$');
-    if (
-      scxmlEnteringStates
-        .get(self)
-        ?.some((id) => id === sanitized || id.endsWith('.' + sanitized))
-    ) {
-      return true;
-    }
-    try {
-      const snap = (self as any).getSnapshot?.();
-      if (snap?.nodes) {
-        return snap.nodes.some(
-          (node: any) =>
-            node.id === sanitized || node.id.endsWith('.' + sanitized)
-        );
-      }
-    } catch {
-      // ignore
-    }
-    return false;
-  };
-  const SCXML_ORIGIN = `#_scxml_${sessionId}`;
-  const SCXML_ORIGIN_TYPE = 'http://www.w3.org/TR/scxml/#SCXMLEventProcessor';
-  const isExternal = event && (event as any)._scxmlExternal;
-  const isInitEvent = event && event.type === '@xstate.init';
-  // Per SCXML spec, _event has these fields. Fields that don't apply to a
-  // particular event are present with value `undefined` (so `'name' in _event`
-  // is true but `typeof _event.name === 'undefined'` for unset fields).
-  // For xstate.done.* events, _event.data is the `output` payload (matches
-  // SCXML's done event semantics where data is the donedata).
-  const isDoneEvent =
-    event &&
-    typeof event.type === 'string' &&
-    event.type.startsWith('xstate.done.');
-  const explicitType =
-    event && ((event as any)._scxmlEventType as string | undefined);
-  let scxmlEvent: object | undefined;
-  if (event && !isInitEvent) {
-    scxmlEvent = scxmlEvents.get(event);
-    if (!scxmlEvent) {
-      scxmlEvent = {
-        name:
-          ((event as any)._scxmlEventName as string | undefined) ??
-          toScxmlEventName(event.type),
-        type: explicitType || (isExternal ? 'external' : 'internal'),
-        sendid: (event as any)._scxmlSendId as string | undefined,
-        origin: isExternal ? SCXML_ORIGIN : undefined,
-        origintype: isExternal ? SCXML_ORIGIN_TYPE : undefined,
-        invokeid:
-          ((event as any)._scxmlInvokeId as string | undefined) ??
-          ((event as any).actorId as string | undefined),
-        data: isDoneEvent
-          ? (event as any).output
-          : (event as any)._scxmlEventData !== undefined
-            ? (event as any)._scxmlEventData
-            : event
-      };
-      scxmlEvents.set(event, scxmlEvent);
-    }
-  }
-  const result = fn(
-    dataContext,
-    scxmlEvent,
-    sessionId,
-    machineName,
-    self
-      ? getScxmlIoProcessors(self, sessionId)
-      : { scxml: { location: `#_scxml_${sessionId}` } },
-    In
-  );
-
-  return result;
-}
-
-/** Executes an SCXML script block and returns updated context values. */
-function executeScript(
-  context: MachineContext,
-  code: string,
-  self: AnyActorRef,
-  global: boolean
-): Record<string, unknown> {
-  // Create a proxy to track which properties are modified
-  const updates: Record<string, unknown> = {};
-  const contextKeys = Object.keys(context);
-  const mutableContextKeys = contextKeys.filter(
-    (key) => !SCXML_SYSTEM_VARIABLES.has(key)
-  );
-
-  // Build variable declarations and reassignment capture
-  const varDeclarations = mutableContextKeys
-    .map((k) => `var ${k} = context.${k};`)
-    .join('\n');
-  const captureUpdates = mutableContextKeys
-    .map((k) => `updates.${k} = ${k};`)
-    .join('\n');
-  const codeWithoutComments = code.replace(/\/\/.*$/gm, '');
-  const declarations = getFunctionDeclarations(code);
-  const codeWithoutFunctions = declarations.reduce(
-    (source, declaration) => source.replace(declaration, ''),
-    codeWithoutComments
-  );
-  const declaredVariables = global
-    ? [
-        ...codeWithoutFunctions.matchAll(/\bvar\s+([$_a-zA-Z][$_a-zA-Z0-9]*)/g)
-      ].map((match) => match[1])
-    : [];
-  const captureVariables = declaredVariables
-    .filter(
-      (name) => !contextKeys.includes(name) && !SCXML_SYSTEM_VARIABLES.has(name)
-    )
-    .map((name) => `updates.${name} = ${name};`)
-    .join('\n');
-  const environment = scxmlScriptEnvironments.get(self) ?? {
-    declarations: []
-  };
-  scxmlScriptEnvironments.set(self, environment);
-  for (const declaration of declarations) {
-    if (!environment.declarations.includes(declaration)) {
-      environment.declarations.push(declaration);
-    }
-  }
-
-  const fnBody = `
-${varDeclarations}
-${environment.declarations.join('\n')}
-${code}
-${captureUpdates}
-${captureVariables}
-return updates;
-  `;
-  const sessionId = (self as any).sessionId;
-  const machineName = scxmlMachineNames.get(self) ?? '(machine)';
-  // eslint-disable-next-line @typescript-eslint/no-implied-eval
-  const fn = new Function(
-    'context',
-    'updates',
-    '_sessionid',
-    '_name',
-    '_ioprocessors',
-    fnBody
-  );
-  return fn(
-    context,
-    updates,
-    sessionId,
-    machineName,
-    getScxmlIoProcessors(self, sessionId)
-  );
-}
-
-function createScxmlDomElement(element: ScxmlXmlJSON): any {
-  const children = (element.children ?? []).map(createScxmlDomElement);
-  return {
-    tagName: element.name,
-    getAttribute: (name: string) => element.attributes?.[name] ?? null,
-    getElementsByTagName: (name: string): any[] => [
-      ...(element.name === name ? [createScxmlDomElement(element)] : []),
-      ...children.flatMap((child: any) => child.getElementsByTagName(name))
-    ]
-  };
-}
-
 export function createMachineFromConfig(
   json: MachineJSON,
   sources: MachineSources = {}
@@ -1094,7 +652,6 @@ export function createMachineFromConfig(
   );
   const { evaluateResolvable, resolveValue, makeScope, getDurationConfig } =
     expressionResolver;
-  const sendCounters = new WeakMap<AnyActorRef, number>();
 
   type ResolvedCondition = ((args: any) => boolean) | undefined;
 
@@ -1103,9 +660,7 @@ export function createMachineFromConfig(
     slot: 'guard' | 'choice',
     path: string
   ): ResolvedCondition {
-    if (!condition) {
-      return undefined;
-    }
+    if (!condition) return undefined;
     if (isResolvable(condition)) {
       return (args: any) =>
         !!evaluateResolvable(condition, slot, makeScope(args), path);
@@ -1136,10 +691,16 @@ export function createMachineFromConfig(
     };
   }
 
+  function getMissingGuardMessage(type: string, guards: Record<string, any>) {
+    return `Guard '${type}' is not implemented in machine '${json.id ?? '(machine)'}'. Available guards: ${
+      Object.keys(guards)
+        .map((key) => `'${key}'`)
+        .join(', ') || '(none)'
+    }.`;
+  }
+
   function makeChoiceConfig(choice: StateNodeJSON['choice'], path: string) {
-    if (!choice) {
-      return undefined;
-    }
+    if (!choice) return undefined;
     if (isResolvable(choice)) {
       return (args: any) =>
         evaluateResolvable(choice, 'choice', makeScope(args), path);
@@ -1183,464 +744,26 @@ export function createMachineFromConfig(
     };
   }
 
-  // Pending transition actions: set by .to functions, consumed by entry functions.
-  // This bridges SCXML's exit→transition→entry action ordering with XState's
-  // .to function receiving pre-exit context.
-  // Map keyed by target state ID so parallel transitions don't overwrite each other.
-  const pendingTransitionActionsMap: Record<string, ActionJSON[]> = {};
-
-  // Ordered queue of ALL transition actions (targeted + targetless) for parallel
-  // context sharing. In SCXML, all transition actions execute sequentially in
-  // document order with a shared evolving data model.
-  const allTransitionActions: ActionJSON[][] = [];
-  // Pre-transition context saved when a targetless .to executes. Used by entry
-  // functions to re-execute all transition actions from scratch when parallel
-  // targetless transitions coexist with targeted transitions.
-  let contextBeforeTargetless: MachineContext | null = null;
-  let targetlessEvent: AnyEventObject | null = null;
-  // Platform errors collected during transition selection (e.g., failing
-  // <transition cond>). Drained by the next state's entry into the internal
-  // event queue so they're processed before any external events.
-  const pendingPlatformErrors: Array<Record<string, unknown>> = [];
-  const scxmlDonedataValues = new WeakMap<AnyActorRef, Map<string, unknown>>();
-
-  function evaluateDonedataValue(
-    donedata: ScxmlDonedataJSON,
-    context: MachineContext,
-    event: AnyEventObject,
-    self: AnyActorRef
-  ): unknown {
-    if (donedata.params?.length) {
-      const out: Record<string, unknown> = {};
-      for (const param of donedata.params) {
-        out[param.name] = evaluateExpr(context, param.expr, event, self);
-      }
-      return out;
-    }
-    if (donedata.contentExpr !== undefined) {
-      return evaluateExpr(context, donedata.contentExpr, event, self);
-    }
-    if (donedata.contentText !== undefined) {
-      try {
-        return evaluateExpr(context, donedata.contentText, event, self);
-      } catch {
-        return donedata.contentText;
-      }
-    }
-    return undefined;
-  }
-
-  function getMissingGuardMessage(type: string, guards: Record<string, any>) {
-    return `Guard '${type}' is not implemented in machine '${json.id ?? '(machine)'}'. Available guards: ${
-      Object.keys(guards)
-        .map((key) => `'${key}'`)
-        .join(', ') || '(none)'
-    }.`;
-  }
-
   function resolveRouteConfig(route: StateNodeJSON['route'], path: string) {
-    if (!route || typeof route === 'function') {
-      return route;
-    }
+    if (!route || typeof route === 'function') return route;
     if (isResolvable(route)) {
       return (args: any) =>
         evaluateResolvable(route, 'transition', makeScope(args), path);
     }
     const { guard, ...routeConfig } = route;
-    if (!guard) {
-      return routeConfig;
-    }
+    if (!guard) return routeConfig;
     const resolvedGuard = resolveCondition({ type: guard }, 'guard', path);
     return (args: any) => (resolvedGuard!(args) ? routeConfig : undefined);
   }
 
-  function iterNode(
-    node: StateNodeJSON,
-    nodeKey?: string,
-    ancestorIds: string[] = []
-  ) {
-    const originalEntryActions: ActionJSON[] = [
-      ...(node._scxmlData?.length
-        ? [{ type: 'scxml.datamodel' as const, data: node._scxmlData }]
-        : []),
-      ...toActionArray(node.entry)
-    ];
-    const stateId = node.id || nodeKey;
-
-    // Wrap entry to first execute any pending transition actions, then normal entry.
-    // This ensures SCXML execution order: exit → transition_actions → entry_actions
-    // because pending transition actions are set by .to (before exit) but executed
-    // in entry (after exit), thus seeing post-exit context.
-    const entryFn: Action<any, any, any, any, any, any, any> | undefined = (
-      x,
-      enq
-    ) => {
-      scxmlMachineNames.set(x.self, json.id ?? '(machine)');
-      if (stateId) {
-        scxmlEnteringStates.set(x.self, [...ancestorIds, stateId]);
-      }
-      // Drain any platform errors queued during transition selection (e.g.,
-      // failing <transition cond="..."> evaluations). These must be raised
-      // internally so they're processed before subsequent external events.
-      while (pendingPlatformErrors.length) {
-        const err = pendingPlatformErrors.shift()!;
-        enq.raise({
-          type: 'xstate.error.execution',
-          error: err,
-          _scxmlEventType: 'platform',
-          _scxmlEventName: 'error.execution',
-          _scxmlEventData: err
-        } as AnyEventObject);
-      }
-
-      let context: MachineContext | undefined;
-
-      // If targetless transitions were interleaved with targeted transitions
-      // (parallel state), re-execute ALL transition actions from the original
-      // pre-transition context. This overrides any .to-produced context with
-      // the correct SCXML document-order sequential execution.
-      // Only trigger when BOTH targeted (pending in map) and targetless fired
-      // in the same microstep — prevents stale data from previous events.
-      if (
-        contextBeforeTargetless &&
-        targetlessEvent === x.event &&
-        allTransitionActions.length > 0 &&
-        Object.keys(pendingTransitionActionsMap).length > 0
-      ) {
-        let ctx = contextBeforeTargetless;
-        for (const actions of allTransitionActions) {
-          const mergedX = { ...x, context: ctx };
-          const result = executeActions(actions, mergedX, enq);
-          if (result.context) {
-            ctx = result.context;
-          }
-        }
-        allTransitionActions.length = 0;
-        contextBeforeTargetless = null;
-        targetlessEvent = null;
-        // Clear per-target map since we re-processed everything
-        for (const key of Object.keys(pendingTransitionActionsMap)) {
-          delete pendingTransitionActionsMap[key];
-        }
-        context = ctx;
-      } else {
-        // Normal path: consume pending transition actions for THIS state.
-        // In parallel states, each target gets its own pending actions.
-        const transActions = stateId
-          ? pendingTransitionActionsMap[stateId]
-          : undefined;
-        if (transActions) {
-          delete pendingTransitionActionsMap[stateId!];
-          const result = executeActions(transActions, x, enq);
-          if (result.context) {
-            context = result.context;
-          }
-        }
-        // Clear stale targetless data from previous microsteps
-        contextBeforeTargetless = null;
-        targetlessEvent = null;
-        allTransitionActions.length = 0;
-      }
-
-      // Execute normal entry actions
-      if (originalEntryActions.length) {
-        const mergedX = context ? { ...x, context } : x;
-        const result = executeActions(originalEntryActions, mergedX, enq);
-        if (result.context) {
-          context = result.context;
-        }
-      }
-
-      if (node._scxmlDonedata && stateId) {
-        let values = scxmlDonedataValues.get(x.self);
-        if (!values) {
-          values = new Map();
-          scxmlDonedataValues.set(x.self, values);
-        }
-        try {
-          values.set(
-            stateId,
-            evaluateDonedataValue(
-              node._scxmlDonedata,
-              context ?? x.context,
-              x.event,
-              x.self
-            )
-          );
-        } catch (err) {
-          values.set(stateId, undefined);
-          raiseErrorExecution(
-            { enq, self: x.self, errored: false },
-            'donedata',
-            err
-          );
-        }
-      }
-
-      scxmlEnteringStates.delete(x.self);
-      return { context };
-    };
-
-    const nodeConfig: any = {
-      id: node.id,
-      initial: node._scxmlInitial
-        ? {
-            target: node._scxmlInitial.targets,
-            to: node._scxmlInitial.actions
-              ? (x: any, enq: any) => ({
-                  ...executeActions(node._scxmlInitial!.actions!, x, enq),
-                  target: node._scxmlInitial!.targets
-                })
-              : undefined
-          }
-        : node.initial,
-      type: node.type,
-      history: node.history,
-      target: node.target,
-      _scxmlHistoryActions: node._scxmlHistoryActions
-        ? (x: any, enq: any) => ({
-            ...executeActions(node._scxmlHistoryActions!, x, enq),
-            target: node.target
-          })
-        : undefined,
-      description: node.description,
-      tags: node.tags,
-      input: node.input,
-      timeout: getDurationConfig(node.timeout, `$.states.${nodeKey}.timeout`),
-      states: node.states
-        ? Object.entries(node.states).reduce(
-            (acc, [key, value]) => {
-              acc[key] = iterNode(
-                value,
-                key,
-                stateId ? [...ancestorIds, stateId] : ancestorIds
-              );
-              return acc;
-            },
-            {} as Record<
-              string,
-              Next_StateNodeConfig<
-                any,
-                any,
-                any,
-                any,
-                any,
-                any,
-                any,
-                any,
-                any,
-                any,
-                any,
-                any
-              >
-            >
-          )
-        : undefined,
-      on: node.on
-        ? Object.entries(node.on).reduce(
-            (acc, [key, value]) => {
-              acc[key] = getTransitionConfig(value);
-              return acc;
-            },
-            {} as Record<string, any>
-          )
-        : undefined,
-      always: node.always ? getTransitionConfig(node.always) : undefined,
-      onError: node.onError ? getTransitionConfig(node.onError) : undefined,
-      choice: makeChoiceConfig(node.choice, `$.states.${nodeKey}.choice`),
-      route: resolveRouteConfig(node.route, `$.states.${nodeKey}.route`),
-      after: node.after
-        ? Object.entries(node.after).reduce(
-            (acc, [key, value]) => {
-              acc[key] = getTransitionConfig(value);
-              return acc;
-            },
-            {} as Record<string, any>
-          )
-        : undefined,
-      onTimeout: node.onTimeout
-        ? getTransitionConfig(node.onTimeout)
-        : undefined,
-      entry: node.type === 'choice' ? undefined : (entryFn as any),
-      exit: node.exit
-        ? (iterActions(toActionArray(node.exit)) as any)
-        : undefined,
-      invoke: node.invoke ? iterInvokeConfigs(node.invoke) : undefined,
-      meta: node.meta,
-      output: node._scxmlDonedata
-        ? makeDonedataOutput(stateId!)
-        : isResolvable(node.output)
-          ? ({ context, event, self }: any) =>
-              evaluateResolvable(
-                node.output as ResolvableJSON,
-                'output',
-                { context, event, self },
-                `$.states.${nodeKey}.output`
-              )
-          : node.output
-    };
-
-    return nodeConfig;
-  }
-
-  function makeDonedataOutput(stateId: string) {
-    return ({ self }: any) => scxmlDonedataValues.get(self)?.get(stateId);
-  }
-
-  function iterInvokeConfigs(invokes: InvokeJSON | InvokeJSON[]): any {
-    const invokeArray = Array.isArray(invokes) ? invokes : [invokes];
-    return invokeArray.map((inv) => {
-      const extInv = inv as InvokeJSON & { _nestedMachineJSON?: MachineJSON };
-      // Create child machine from nested SCXML JSON
-      let src: any;
-      if (extInv._nestedMachineJSON) {
-        src = createMachineFromConfig(
-          extInv._nestedMachineJSON,
-          resolvedSources
-        );
-      } else {
-        src = resolvedSources.actors[inv.src] ?? inv.src;
-      }
-      return {
-        src,
-        id: inv.id,
-        registryKey: inv.registryKey,
-        input: inv._scxmlInput
-          ? ({ context, event, self }: any) => {
-              const input: Record<string, unknown> = {};
-              for (const name of inv._scxmlInput!.namelist ?? []) {
-                if (!(name in context)) {
-                  throw new ReferenceError(`${name} is not defined`);
-                }
-                input[name] = context[name];
-              }
-              for (const param of inv._scxmlInput!.params ?? []) {
-                input[param.name] = param.expr
-                  ? evaluateExpr(context, param.expr, event, self)
-                  : param.location
-                    ? evaluateExpr(context, param.location, event, self)
-                    : undefined;
-              }
-              return input;
-            }
-          : inv.input !== undefined
-            ? (args: any) =>
-                resolveValue(
-                  inv.input,
-                  'input',
-                  makeScope(args),
-                  '$.invoke.input'
-                )
-            : undefined,
-        onDone: inv.onDone ? getTransitionConfig(inv.onDone) : undefined,
-        onError: inv.onError ? getTransitionConfig(inv.onError) : undefined,
-        onSnapshot: inv.onSnapshot
-          ? getTransitionConfig(inv.onSnapshot)
-          : undefined,
-        timeout: getDurationConfig(inv.timeout, '$.invoke.timeout'),
-        onTimeout: inv.onTimeout
-          ? getTransitionConfig(inv.onTimeout)
-          : undefined
-      };
-    });
-  }
-
-  /**
-   * Raise an SCXML error.execution platform event onto the internal queue. Per
-   * spec, error.execution is type='platform' and _event.data contains error
-   * info. The surrounding block of executable content stops.
-   */
-  function raiseErrorExecution(
-    state: ExecState,
-    tagname: string,
-    err: unknown,
-    sendId?: string
-  ) {
-    const message =
-      err instanceof Error
-        ? err.message
-        : typeof err === 'string'
-          ? err
-          : 'unknown error';
-    const errorEvent = {
-      type: 'xstate.error.execution',
-      error: {
-        tagname,
-        message,
-        line: 0,
-        column: 0,
-        reason: message
-      },
-      _scxmlEventType: 'platform',
-      _scxmlEventName: 'error.execution',
-      ...(sendId ? { _scxmlSendId: sendId } : {}),
-      _scxmlEventData: {
-        tagname,
-        message,
-        line: 0,
-        column: 0,
-        reason: message
-      }
-    } as AnyEventObject;
-    state.enq.raise(errorEvent);
-    state.errored = true;
-  }
-
-  interface ExecState {
-    enq: any;
-    self: AnyActorRef;
-    errored: boolean;
-  }
-
-  function executeActionDefinition(
-    action: CustomActionJSON,
-    x: any,
-    enq: any,
-    params: unknown,
-    parentState: ExecState,
-    stack: string[]
-  ): { context: MachineContext | undefined; errored: boolean } {
-    const definition = json.actions?.[action.type];
-    if (!definition) {
-      enq(x.actions[action.type], params);
-      return { context: undefined, errored: parentState.errored };
-    }
-    if (stack.includes(action.type)) {
-      raiseErrorExecution(
-        parentState,
-        'action',
-        new Error(
-          `Circular action reference: ${stack.concat(action.type).join(' -> ')}`
-        )
-      );
-      return { context: undefined, errored: true };
-    }
-    const definitions = Array.isArray(definition) ? definition : [definition];
-    return executeActions(
-      definitions,
-      { ...x, params },
-      enq,
-      parentState,
-      stack.concat(action.type)
-    );
-  }
-
-  /** Execute an array of SCXML action JSON descriptors with context and enqueue. */
   function executeActions(
     actions: ActionJSON[],
     x: any,
     enq: any,
-    parentState?: ExecState,
     stack: string[] = []
-  ): { context: MachineContext | undefined; errored: boolean } {
-    const state: ExecState = parentState ?? {
-      enq,
-      self: x.self,
-      errored: false
-    };
+  ) {
     let context: MachineContext | undefined;
     for (const action of actions) {
-      if (state.errored) break;
       if (isResolvable(action)) {
         const result = evaluateResolvable(
           action,
@@ -1652,42 +775,26 @@ export function createMachineFromConfig(
           result &&
           typeof result === 'object' &&
           'context' in result &&
-          (result as { context?: MachineContext }).context
+          (result as any).context
         ) {
           context ??= {};
-          Object.assign(
-            context,
-            (result as { context: MachineContext }).context
-          );
+          Object.assign(context, (result as any).context);
         }
-      } else if (isBuiltInActionJSON(action)) {
+        continue;
+      }
+      if (isBuiltInActionJSON(action)) {
         switch (action.type) {
-          case '@xstate.raise': {
-            // Tag as external if it has a delay (from <send>, not <raise>).
-            // Attach _scxmlSendId so _event.sendid can be read in expressions.
-            const isExternal = action.delay !== undefined;
-            const resolvedEvent = resolveValue(
-              action.event,
-              'actionParams',
-              makeScope(x, { params: x.params }),
-              '$.actions.event'
-            ) as EventObject;
-            const event =
-              isExternal || action.id !== undefined
-                ? {
-                    ...resolvedEvent,
-                    ...(isExternal ? { _scxmlExternal: true } : {}),
-                    ...(action.id !== undefined
-                      ? { _scxmlSendId: action.id }
-                      : {})
-                  }
-                : resolvedEvent;
-            enq.raise(event, {
-              id: action.id,
-              delay: action.delay
-            });
+          case '@xstate.raise':
+            enq.raise(
+              resolveValue(
+                action.event,
+                'actionParams',
+                makeScope(x, { params: x.params }),
+                '$.actions.event'
+              ),
+              { id: action.id, delay: action.delay }
+            );
             break;
-          }
           case '@xstate.cancel':
             enq.cancel(action.id);
             break;
@@ -1716,772 +823,207 @@ export function createMachineFromConfig(
               )
             );
             break;
-          default:
-            throw new Error(`Unknown built-in action: ${(action as any).type}`);
         }
-      } else if (action.type === 'scxml.finalize') {
-        const finalize = action as ScxmlFinalizeJSON;
-        const invokeId =
-          (x.event as any)._scxmlInvokeId ?? (x.event as any).actorId;
-        if (invokeId === finalize.invokeId) {
-          const result = executeActions(
-            finalize.actions,
-            { ...x, context: { ...x.context, ...context } },
-            enq,
-            state
-          );
-          if (result.context) {
-            context = result.context;
-          }
-        }
-      } else if (action.type === 'scxml.forward') {
-        const child = x.children?.[(action as ScxmlForwardJSON).invokeId];
-        if (child) {
-          enq.sendTo(child, {
-            ...x.event,
-            _scxmlExternal: true
-          });
-        }
-      } else if (action.type === 'scxml.log') {
-        const scxmlAction = action as ScxmlLogJSON;
-        try {
-          const value = scxmlAction.expr
-            ? evaluateExpr(
-                { ...x.context, ...context },
-                scxmlAction.expr,
-                x.event,
-                x.self
-              )
-            : undefined;
-          enq.log(
-            ...(scxmlAction.label ? [scxmlAction.label, value] : [value])
-          );
-        } catch (err) {
-          raiseErrorExecution(state, 'log', err);
-        }
-      } else if (action.type === 'scxml.assign') {
-        const scxmlAction = action as ScxmlAssignJSON;
-        const mergedContext = { ...x.context, ...context };
-        let value: unknown;
-        try {
-          value = evaluateExpr(
-            mergedContext,
-            scxmlAction.expr,
-            x.event,
-            x.self
-          );
-        } catch (err) {
-          raiseErrorExecution(state, 'assign', err);
-          continue;
-        }
-        if (SCXML_SYSTEM_VARIABLES.has(scxmlAction.location.split('.')[0])) {
-          raiseErrorExecution(
-            state,
-            'assign',
-            new Error(
-              `Cannot assign to SCXML system variable: ${scxmlAction.location}`
-            )
-          );
-          continue;
-        }
-        if (scxmlAction.location.includes('.')) {
-          const path = scxmlAction.location.split('.');
-          const root = path.shift()!;
-          let target = mergedContext[root];
-          if (!target || typeof target !== 'object') {
-            raiseErrorExecution(
-              state,
-              'assign',
-              new Error(`Invalid assign location: ${scxmlAction.location}`)
-            );
-            continue;
-          }
-          const clonedRoot = Array.isArray(target)
-            ? [...target]
-            : { ...target };
-          target = clonedRoot;
-          for (const segment of path.slice(0, -1)) {
-            const child = (target as Record<string, unknown>)[segment];
-            if (!child || typeof child !== 'object') {
-              raiseErrorExecution(
-                state,
-                'assign',
-                new Error(`Invalid assign location: ${scxmlAction.location}`)
-              );
-              break;
-            }
-            const clonedChild = Array.isArray(child)
-              ? [...child]
-              : { ...child };
-            (target as Record<string, unknown>)[segment] = clonedChild;
-            target = clonedChild;
-          }
-          if (state.errored) continue;
-          (target as Record<string, unknown>)[path.at(-1)!] = value;
-          context ??= {};
-          context[root] = clonedRoot;
-          continue;
-        }
-        context ??= {};
-        context[scxmlAction.location] = value;
-      } else if (action.type === 'scxml.datamodel') {
-        const scxmlAction = action as ScxmlDatamodelJSON;
-        context ??= {};
-        for (const data of scxmlAction.data) {
-          if (
-            Array.isArray(x.context._xstateScxmlInputKeys) &&
-            x.context._xstateScxmlInputKeys.includes(data.id)
-          ) {
-            continue;
-          }
-          if (data.src) {
-            raiseErrorExecution(
-              state,
-              'data',
-              new Error(`Unresolved data resource: ${data.src}`)
-            );
-            break;
-          }
-          if (data.expr !== undefined) {
-            try {
-              context[data.id] = evaluateExpr(
-                { ...x.context, ...context },
-                data.expr,
-                x.event,
-                x.self
-              );
-            } catch (err) {
-              raiseErrorExecution(state, 'data', err);
-              break;
-            }
-          } else if (data.content !== undefined) {
-            try {
-              context[data.id] = evaluateExpr(
-                { ...x.context, ...context },
-                data.content,
-                x.event,
-                x.self
-              );
-            } catch {
-              context[data.id] = data.content.replace(/\s+/g, ' ').trim();
-            }
-          } else if (data.xml) {
-            context[data.id] = createScxmlDomElement(data.xml);
-          } else {
-            if (x.context[data.id] === undefined) {
-              context[data.id] = undefined;
-            }
-          }
-        }
-      } else if (action.type === 'scxml.raise') {
-        const scxmlAction = action as ScxmlRaiseJSON;
-        const mergedContext = { ...x.context, ...context };
-
-        let eventType: string;
-        let eventData: Record<string, unknown>;
-        let target: string | undefined;
-        let delay: number | undefined;
-        try {
-          if (
-            scxmlAction.processorType &&
-            ![
-              'scxml',
-              'http://www.w3.org/TR/scxml/#SCXMLEventProcessor'
-            ].includes(scxmlAction.processorType)
-          ) {
-            throw new Error(
-              `Unsupported send type: ${scxmlAction.processorType}`
-            );
-          }
-          eventType = scxmlAction.eventexpr
-            ? (evaluateExpr(
-                mergedContext,
-                scxmlAction.eventexpr,
-                x.event,
-                x.self
-              ) as string)
-            : scxmlAction.event || 'unknown';
-          eventType = normalizeScxmlEventType(eventType);
-
-          eventData = {
-            type: eventType,
-            _scxmlEventName: toScxmlEventName(eventType)
-          };
-          const payload: Record<string, unknown> = {};
-          if (scxmlAction.namelist) {
-            for (const name of scxmlAction.namelist) {
-              payload[name] = evaluateExpr(
-                mergedContext,
-                name,
-                x.event,
-                x.self
-              );
-            }
-          }
-          if (scxmlAction.params) {
-            for (const param of scxmlAction.params) {
-              payload[param.name] = evaluateExpr(
-                mergedContext,
-                param.expr ?? param.location!,
-                x.event,
-                x.self
-              );
-            }
-          }
-          if (scxmlAction.content) {
-            const content = scxmlAction.content;
-            let contentValue: unknown;
-            if (content.expr !== undefined) {
-              contentValue = evaluateExpr(
-                mergedContext,
-                content.expr,
-                x.event,
-                x.self
-              );
-            } else if (content.xml) {
-              contentValue = createScxmlDomElement(content.xml);
-            } else {
-              const text = content.text?.replace(/\s+/g, ' ').trim() ?? '';
-              try {
-                contentValue = evaluateExpr(
-                  mergedContext,
-                  text,
-                  x.event,
-                  x.self
-                );
-              } catch {
-                contentValue = text;
-              }
-            }
-            eventData._scxmlEventData = contentValue;
-          } else if (Object.keys(payload).length) {
-            Object.assign(eventData, payload);
-            eventData._scxmlEventData = payload;
-          }
-
-          if (scxmlAction.idlocation) {
-            const nextId = (sendCounters.get(x.self) ?? 0) + 1;
-            sendCounters.set(x.self, nextId);
-            const generatedId = `${x.self.sessionId}.send.${nextId}`;
-            context ??= {};
-            context[scxmlAction.idlocation] = generatedId;
-            eventData._scxmlSendId = generatedId;
-          }
-
-          target = scxmlAction.targetexpr
-            ? (evaluateExpr(
-                mergedContext,
-                scxmlAction.targetexpr,
-                x.event,
-                x.self
-              ) as string)
-            : scxmlAction.target;
-
-          const isInternalTarget = target === '#_internal';
-          delay = scxmlAction.delayexpr
-            ? delayToMs(
-                evaluateExpr(
-                  mergedContext,
-                  scxmlAction.delayexpr,
-                  x.event,
-                  x.self
-                ) as string | number
-              )
-            : isInternalTarget
-              ? undefined
-              : scxmlAction.delay;
-        } catch (err) {
-          raiseErrorExecution(state, 'send', err);
-          continue;
-        }
-
-        const isInternalTarget = target === '#_internal';
-        const isParentTarget = target === '#_parent';
-        if (!isInternalTarget) {
-          (eventData as any)._scxmlExternal = true;
-        }
-        if (isParentTarget) {
-          (eventData as any)._scxmlInvokeId = x.self.id;
-        }
-
-        // Validate target. SCXML targets must be '#_internal', '#_parent',
-        // or '#_<id>'. A bare string without the '#_' prefix is invalid and
-        // raises error.execution.
-        if (
-          typeof target === 'string' &&
-          target.length > 0 &&
-          !target.startsWith('#_')
-        ) {
-          raiseErrorExecution(
-            state,
-            'send',
-            new Error(`Invalid send target: ${target}`),
-            eventData._scxmlSendId as string | undefined
-          );
-          continue;
-        }
-        // Attach send id so _event.sendid can be read in expressions.
-        if (
-          scxmlAction.id !== undefined &&
-          eventData._scxmlSendId === undefined
-        ) {
-          (eventData as any)._scxmlSendId = scxmlAction.id;
-        }
-        // Resolve target at runtime: parent, child, or self
-        if (isParentTarget && x.parent) {
-          // Send to parent via sendTo; pass delay if present
-          enq.sendTo(
-            x.parent,
-            eventData as AnyEventObject,
-            delay !== undefined ? { delay } : undefined
-          );
-        } else if (
-          typeof target === 'string' &&
-          target.startsWith('#_') &&
-          !isParentTarget &&
-          !isInternalTarget
-        ) {
-          // #_<invokeId> → try to send to child actor, fall back to self-raise
-          const childId = target.slice(2); // strip '#_'
-          const childRef = x.children?.[childId];
-          if (childRef) {
-            enq.sendTo(childRef, eventData as AnyEventObject);
-          } else if (
-            childId.startsWith('scxml_') &&
-            childId !== `scxml_${x.self.sessionId}`
-          ) {
-            const message = `Unable to dispatch event to target: ${target}`;
-            enq.raise({
-              type: 'xstate.error.communication',
-              error: {
-                tagname: 'send',
-                message,
-                line: NaN,
-                column: NaN,
-                reason: message
-              },
-              _scxmlEventType: 'platform',
-              _scxmlEventName: 'error.communication',
-              _scxmlEventData: {
-                tagname: 'send',
-                message,
-                line: NaN,
-                column: NaN,
-                reason: message
-              }
-            } as AnyEventObject);
-          } else {
-            // This SCXML session id uses the external event queue.
-            (eventData as any)._scxmlExternal = true;
-            enq.sendTo(x.self, eventData as AnyEventObject, {
-              id: scxmlAction.id,
-              delay
-            });
-          }
-        } else {
-          // #_internal uses the internal queue; self-send/no target use the
-          // external event queue.
-          if (isInternalTarget) {
-            enq.raise(eventData as AnyEventObject, {
-              id: scxmlAction.id,
-              delay
-            });
-          } else {
-            (eventData as any)._scxmlExternal = true;
-            enq.sendTo(x.self, eventData as AnyEventObject, {
-              id: scxmlAction.id,
-              delay
-            });
-          }
-        }
-      } else if (action.type === 'scxml.cancel') {
-        const scxmlAction = action as ScxmlCancelJSON;
-        const mergedContext = { ...x.context, ...context };
-        try {
-          const sendId = evaluateExpr(
-            mergedContext,
-            scxmlAction.sendidexpr,
-            x.event,
-            x.self
-          ) as string;
-          if (sendId) {
-            enq.cancel(sendId);
-          }
-        } catch {
-          // ignore evaluation errors
-        }
-      } else if (action.type === 'scxml.script') {
-        const scxmlAction = action as ScxmlScriptJSON;
-        const mergedContext = { ...x.context, ...context };
-        try {
-          const updatedContext = executeScript(
-            mergedContext,
-            scxmlAction.code,
-            x.self,
-            scxmlAction.global ?? false
-          );
-          context ??= {};
-          Object.assign(context, updatedContext);
-        } catch (err) {
-          raiseErrorExecution(state, 'script', err);
-        }
-      } else if (action.type === 'scxml.foreach') {
-        const scxmlAction = action as ScxmlForeachJSON;
-        const mergedContext = { ...x.context, ...context };
-        // SCXML: `item` (and `index` if present) must be valid datamodel
-        // locations — i.e. legal identifiers. Reject things like 'continue'
-        // (string literal) or '7' (number).
-        const isLegalIdentifier = (s: string) =>
-          /^[$_a-zA-Z][$_a-zA-Z0-9]*$/.test(s);
-        if (
-          !isLegalIdentifier(scxmlAction.item) ||
-          (scxmlAction.index && !isLegalIdentifier(scxmlAction.index))
-        ) {
-          raiseErrorExecution(
-            state,
-            'foreach',
-            new Error('foreach item/index is not a legal location')
-          );
-          continue;
-        }
-        let arr: unknown;
-        try {
-          arr = evaluateExpr(mergedContext, scxmlAction.array, x.event, x.self);
-        } catch (err) {
-          raiseErrorExecution(state, 'foreach', err);
-          continue;
-        }
-        if (!Array.isArray(arr)) {
-          raiseErrorExecution(
-            state,
-            'foreach',
-            new Error('foreach array is not iterable')
-          );
-          continue;
-        }
-        context ??= {};
-        for (let i = 0; i < arr.length; i++) {
-          context[scxmlAction.item] = arr[i];
-          if (scxmlAction.index) {
-            context[scxmlAction.index] = i;
-          }
-          const iterX = { ...x, context: { ...x.context, ...context } };
-          const iterResult = executeActions(
-            scxmlAction.actions,
-            iterX,
-            enq,
-            state
-          );
-          if (iterResult.context) {
-            Object.assign(context, iterResult.context);
-          }
-          if (state.errored) break;
-        }
-      } else if (action.type === 'scxml.block') {
-        const scxmlAction = action as ScxmlBlockJSON;
-        const blockX = { ...x, context: { ...x.context, ...context } };
-        // Block has its own ExecState — errors don't propagate to parent.
-        const blockResult = executeActions(scxmlAction.actions, blockX, enq);
-        if (blockResult.context) {
-          context ??= {};
-          for (const key of Object.keys(blockResult.context)) {
-            if (blockResult.context[key] !== x.context[key]) {
-              context[key] = blockResult.context[key];
-            }
-          }
-        }
-      } else if (action.type === 'scxml.if') {
-        const scxmlAction = action as ScxmlIfJSON;
-        const mergedContext = { ...x.context, ...context };
-        for (const branch of scxmlAction.branches) {
-          let condResult: boolean;
-          if (branch.cond) {
-            try {
-              condResult = !!evaluateExpr(
-                mergedContext,
-                branch.cond,
-                x.event,
-                x.self
-              );
-            } catch (err) {
-              raiseErrorExecution(state, 'if', err);
-              condResult = false;
-            }
-          } else {
-            condResult = true;
-          }
-          if (condResult) {
-            if (branch.actions.length) {
-              const branchX = { ...x, context: mergedContext };
-              const branchResult = executeActions(
-                branch.actions,
-                branchX,
-                enq,
-                state
-              );
-              if (branchResult.context) {
-                context ??= {};
-                for (const key of Object.keys(branchResult.context)) {
-                  if (branchResult.context[key] !== x.context[key]) {
-                    context[key] = branchResult.context[key];
-                  }
-                }
-              }
-            }
-            break;
-          }
-        }
-      } else {
-        const params = resolveValue(
-          (action as CustomActionJSON).params,
-          'actionParams',
-          makeScope(x, { params: x.params }),
-          '$.actions.params'
-        );
-        const result = executeActionDefinition(
-          action,
-          x,
-          enq,
-          params,
-          state,
-          stack
-        );
-        if (result.context) {
-          context ??= {};
-          Object.assign(context, result.context);
-        }
+        continue;
       }
-    }
-    return {
-      context: context ? { ...x.context, ...context } : undefined,
-      errored: state.errored
-    };
-  }
-
-  function iterActions(
-    actions: ActionJSON[]
-  ): Action<any, any, any, any, any, any, any> {
-    return (x, enq) => executeActions(actions, x, enq);
-  }
-
-  function getFinalizedGuardContext(
-    context: MachineContext,
-    event: AnyEventObject,
-    self: AnyActorRef,
-    finalizers: NonNullable<TransitionJSON['_scxml']>['finalize']
-  ): MachineContext {
-    const invokeId = (event as any)._scxmlInvokeId ?? (event as any).actorId;
-    const finalized = { ...context };
-    for (const finalizer of finalizers ?? []) {
-      if (finalizer.invokeId !== invokeId) continue;
-      for (const action of finalizer.actions) {
-        if (!('type' in action) || action.type !== 'scxml.assign') continue;
-        const assignAction = action as ScxmlAssignJSON;
-        finalized[assignAction.location] = evaluateExpr(
-          finalized,
-          assignAction.expr,
-          event,
-          self
+      const params = resolveValue(
+        action.params,
+        'actionParams',
+        makeScope(x, { params: x.params }),
+        '$.actions.params'
+      );
+      const definition = json.actions?.[action.type];
+      if (!definition) {
+        enq(x.actions[action.type], params);
+        continue;
+      }
+      if (stack.includes(action.type)) {
+        throw new Error(
+          `Circular action reference: ${stack.concat(action.type).join(' -> ')}`
         );
       }
+      const definitions = Array.isArray(definition) ? definition : [definition];
+      const result = executeActions(
+        definitions,
+        { ...x, context: { ...x.context, ...context }, params },
+        enq,
+        stack.concat(action.type)
+      );
+      if (result.context) context = result.context;
     }
-    return finalized;
+    return { context: context ? { ...x.context, ...context } : undefined };
+  }
+
+  function iterActions(actions: ActionJSON | ActionJSON[]) {
+    const actionArray = toActionArray(actions);
+    return ((x: any, enq: any) =>
+      executeActions(actionArray, x, enq)) as Action<
+      any,
+      any,
+      any,
+      any,
+      any,
+      any,
+      any,
+      any,
+      any
+    >;
   }
 
   function getTransitionConfig(
     transition: TransitionConfigJSON | TransitionConfigJSON[]
   ): any {
     const transitions = Array.isArray(transition) ? transition : [transition];
-
-    // Return an array of transition configs. Each SCXML transition becomes
-    // a separate XState transition with its own guard and optional .to.
-    // This ensures guards are evaluated by XState's evaluateCandidate (once,
-    // with pre-exit context) and NOT re-evaluated in computeEntrySet.
-    return transitions.map((t, index) => {
-      if (isResolvable(t)) {
+    return transitions.map((item, index) => {
+      if (isResolvable(item)) {
         return (x: any, enq: any) =>
           evaluateResolvable(
-            t,
+            item,
             'transition',
             makeScope(x, { enq }),
             `$.transition${transitions.length > 1 ? `[${index}]` : ''}`
           );
       }
-      const target = Array.isArray(t.target) ? t.target[0] : t.target;
-      const targetConfig = t.target;
-      const resolvedGuard = resolveCondition(
-        t.guard,
-        'guard',
-        '$.transition.guard'
-      );
-      const guard =
-        resolvedGuard && t._scxml?.finalize?.length
-          ? (args: any) =>
-              resolvedGuard({
-                ...args,
-                context: getFinalizedGuardContext(
-                  args.context,
-                  args.event,
-                  args.self,
-                  t._scxml!.finalize
-                )
-              })
-          : resolvedGuard;
-      const resolveTransitionContext = (x: any) =>
-        t.context
-          ? (resolveValue(
-              t.context,
+      const context = item.context
+        ? (x: any) =>
+            resolveValue(
+              item.context,
               'transitionContext',
               makeScope(x),
               '$.transition.context'
-            ) as MachineContext)
-          : undefined;
-      const resolveTransitionInput = (x: any) =>
-        t.input !== undefined
-          ? resolveValue(t.input, 'input', makeScope(x), '$.transition.input')
-          : undefined;
-      const eventMatcher = t._scxml?.eventDescriptors
-        ? (event: EventObject) =>
-            t._scxml!.eventDescriptors!.some((descriptor) =>
-              matchesScxmlEventDescriptor(descriptor, event)
-            )
+            ) as MachineContext
         : undefined;
-
-      // No guard and no actions: simple static config
-      if (!guard && !t.actions?.length) {
-        if (t.context) {
-          return {
-            matches: t.matches,
-            _scxml: t._scxml,
-            _eventMatcher: eventMatcher,
-            to: (x: any) => ({
-              target: targetConfig,
-              context: resolveTransitionContext(x),
-              reenter: t.reenter
-            }),
-            description: t.description,
-            meta: t.meta,
-            input: t.input !== undefined ? resolveTransitionInput : undefined
-          };
-        }
-        return {
-          matches: t.matches,
-          _scxml: t._scxml,
-          _eventMatcher: eventMatcher,
-          target: targetConfig,
-          description: t.description,
-          reenter: t.reenter,
-          meta: t.meta,
-          input: t.input !== undefined ? resolveTransitionInput : undefined
-        };
-      }
-
-      // Guard but no actions: static config with guard
-      if (guard && !t.actions?.length) {
-        if (t.context) {
-          return {
-            matches: t.matches,
-            _scxml: t._scxml,
-            _eventMatcher: eventMatcher,
-            guard,
-            to: (x: any) => ({
-              target: targetConfig,
-              context: resolveTransitionContext(x),
-              reenter: t.reenter
-            }),
-            description: t.description,
-            meta: t.meta,
-            input: t.input !== undefined ? resolveTransitionInput : undefined
-          };
-        }
-        return {
-          matches: t.matches,
-          _scxml: t._scxml,
-          _eventMatcher: eventMatcher,
-          target: targetConfig,
-          guard,
-          description: t.description,
-          reenter: t.reenter,
-          meta: t.meta,
-          input: t.input !== undefined ? resolveTransitionInput : undefined
-        };
-      }
-
-      // Targetless transitions: execute actions directly in .to AND track for
-      // parallel re-execution. For non-parallel, .to result is used directly.
-      // For parallel (entries exist), first entry re-executes all in document order.
-      if (!target) {
-        return {
-          matches: t.matches,
-          _scxml: t._scxml,
-          _eventMatcher: eventMatcher,
-          guard,
-          description: t.description,
-          reenter: t.reenter,
-          meta: t.meta,
-          input: t.input !== undefined ? resolveTransitionInput : undefined,
-          to: (x: any, enq: any) => {
-            const context = resolveTransitionContext(x);
-            if (t.actions?.length) {
-              // Track for parallel re-execution (dedup by reference)
-              if (!allTransitionActions.includes(t.actions)) {
-                allTransitionActions.push(t.actions);
-              }
-              // Save pre-transition context for parallel override
-              contextBeforeTargetless ??= x.context;
-              targetlessEvent ??= x.event;
-              // Execute immediately (fallback for non-parallel case)
-              const result = executeActions(t.actions, x, enq);
-              if (result.context) {
-                return { context: { ...context, ...result.context } };
-              }
-            }
-            return context ? { context } : {};
-          }
-        };
-      }
-
-      // Has target + actions: use guard (for XState evaluation) + .to (for pending actions).
-      // The .to function does NOT re-check the guard — XState's evaluateCandidate
-      // already validated it. The .to just stores actions for entry to execute.
-      // Map by target state ID so parallel transitions each get their own actions.
+      const input =
+        item.input !== undefined
+          ? (x: any) =>
+              resolveValue(
+                item.input,
+                'input',
+                makeScope(x),
+                '$.transition.input'
+              )
+          : undefined;
+      const dynamic = !!context || !!item.actions?.length;
       return {
-        matches: t.matches,
-        _scxml: t._scxml,
-        _eventMatcher: eventMatcher,
-        guard,
-        description: t.description,
-        reenter: t.reenter,
-        meta: t.meta,
-        input: t.input !== undefined ? resolveTransitionInput : undefined,
-        to: (_x: any, _enq: any) => {
-          const context = resolveTransitionContext(_x);
-          if (t.actions?.length) {
-            const targetId = target.replace(/^#/, '');
-            pendingTransitionActionsMap[targetId] = t.actions;
-            // Track for parallel re-execution (dedup by reference)
-            if (!allTransitionActions.includes(t.actions)) {
-              allTransitionActions.push(t.actions);
+        matches: item.matches,
+        target: dynamic ? undefined : item.target,
+        to: dynamic
+          ? (x: any, enq: any) => {
+              const resolvedContext = context?.(x);
+              const result = item.actions?.length
+                ? executeActions(
+                    item.actions,
+                    resolvedContext
+                      ? { ...x, context: { ...x.context, ...resolvedContext } }
+                      : x,
+                    enq
+                  )
+                : undefined;
+              return {
+                target: item.target,
+                context: result?.context ?? resolvedContext,
+                reenter: item.reenter
+              };
             }
-          }
-          return {
-            target: targetConfig,
-            context,
-            reenter: t.reenter
-          };
-        }
+          : undefined,
+        guard: resolveCondition(item.guard, 'guard', '$.transition.guard'),
+        description: item.description,
+        reenter: item.reenter,
+        meta: item.meta,
+        input
       };
     });
   }
 
+  function iterInvokeConfigs(invokes: InvokeJSON | InvokeJSON[]): any {
+    return (Array.isArray(invokes) ? invokes : [invokes]).map((inv) => ({
+      src: resolvedSources.actors[inv.src] ?? inv.src,
+      id: inv.id,
+      registryKey: inv.registryKey,
+      input:
+        inv.input !== undefined
+          ? (args: any) =>
+              resolveValue(
+                inv.input,
+                'input',
+                makeScope(args),
+                '$.invoke.input'
+              )
+          : undefined,
+      onDone: inv.onDone ? getTransitionConfig(inv.onDone) : undefined,
+      onError: inv.onError ? getTransitionConfig(inv.onError) : undefined,
+      onSnapshot: inv.onSnapshot
+        ? getTransitionConfig(inv.onSnapshot)
+        : undefined,
+      timeout: getDurationConfig(inv.timeout, '$.invoke.timeout'),
+      onTimeout: inv.onTimeout ? getTransitionConfig(inv.onTimeout) : undefined
+    }));
+  }
+
+  function iterNode(node: StateNodeJSON, nodeKey?: string): any {
+    return {
+      id: node.id,
+      initial: node.initial,
+      type: node.type,
+      history: node.history,
+      target: node.target,
+      description: node.description,
+      tags: node.tags,
+      input: node.input,
+      timeout: getDurationConfig(node.timeout, `$.states.${nodeKey}.timeout`),
+      states: node.states
+        ? Object.fromEntries(
+            Object.entries(node.states).map(([key, value]) => [
+              key,
+              iterNode(value, key)
+            ])
+          )
+        : undefined,
+      on: node.on
+        ? Object.fromEntries(
+            Object.entries(node.on).map(([key, value]) => [
+              key,
+              getTransitionConfig(value)
+            ])
+          )
+        : undefined,
+      always: node.always ? getTransitionConfig(node.always) : undefined,
+      onError: node.onError ? getTransitionConfig(node.onError) : undefined,
+      choice: makeChoiceConfig(node.choice, `$.states.${nodeKey}.choice`),
+      route: resolveRouteConfig(node.route, `$.states.${nodeKey}.route`),
+      after: node.after
+        ? Object.fromEntries(
+            Object.entries(node.after).map(([key, value]) => [
+              key,
+              getTransitionConfig(value)
+            ])
+          )
+        : undefined,
+      onTimeout: node.onTimeout
+        ? getTransitionConfig(node.onTimeout)
+        : undefined,
+      entry: node.entry ? iterActions(node.entry) : undefined,
+      exit: node.exit ? iterActions(node.exit) : undefined,
+      invoke: node.invoke ? iterInvokeConfigs(node.invoke) : undefined,
+      meta: node.meta,
+      output: isResolvable(node.output)
+        ? ({ context, event, self }: any) =>
+            evaluateResolvable(
+              node.output as ResolvableJSON,
+              'output',
+              { context, event, self },
+              `$.states.${nodeKey}.output`
+            )
+        : node.output
+    };
+  }
+
   if (json.delays) {
-    for (const key of Object.keys(json.delays)) {
-      const delay = json.delays[key];
-      if (typeof delay === 'number') {
-        resolvedSources.delays[key] = delay;
-      } else if (typeof delay === 'string') {
+    for (const [key, delay] of Object.entries(json.delays)) {
+      if (typeof delay === 'number') resolvedSources.delays[key] = delay;
+      else if (typeof delay === 'string')
         resolvedSources.delays[key] = delayToMs(delay);
-      } else if (delay && typeof delay === 'object') {
+      else if (delay && typeof delay === 'object') {
         resolvedSources.delays[key] = (args: any) =>
           delayToMs(
             resolveValue(
@@ -2496,103 +1038,49 @@ export function createMachineFromConfig(
   }
 
   assertMachineJSON(json, resolvedSources, expressionResolver);
-
-  const rootNodeConfig = iterNode(json);
-  const contextConfig =
-    json.context || json._scxmlDataIds?.length
-      ? {
-          context: (args: any) => ({
-            ...Object.fromEntries(
-              (json._scxmlDataIds ?? []).map((id) => [id, undefined])
-            ),
-            ...(json.context
-              ? (resolveValue(
-                  json.context,
-                  'context',
-                  args,
-                  '$.context'
-                ) as MachineContext)
-              : {}),
-            ...(json._scxmlDataIds?.length &&
-            args.input &&
-            typeof args.input === 'object'
-              ? args.input
-              : {}),
-            ...(json._scxmlDataIds?.length &&
-            args.input &&
-            typeof args.input === 'object'
-              ? { _xstateScxmlInputKeys: Object.keys(args.input) }
-              : {})
-          })
-        }
-      : {};
-
+  const contextConfig = json.context
+    ? {
+        context: (args: any) =>
+          resolveValue(
+            json.context,
+            'context',
+            args,
+            '$.context'
+          ) as MachineContext
+      }
+    : {};
   const machine = createMachine({
-    ...rootNodeConfig,
+    ...iterNode(json),
     ...contextConfig,
     version: json.version
   }) as unknown as AnyStateMachine;
-
-  // Register SCXML guard sources
-  const providedGuards: Record<string, (args: any, params: any) => boolean> = {
-    'scxml.cond': ({ context, event, self }: any, params: any) => {
-      const expr = params?.expr as string;
-      if (!expr) return true;
-      try {
-        return !!evaluateExpr(context, expr, event, self);
-      } catch (err) {
-        // Per SCXML spec, a cond that fails to evaluate is treated as false
-        // AND raises error.execution. We can't enqueue from a guard, so
-        // queue it for the next entry to drain.
-        const message =
-          err instanceof Error
-            ? err.message
-            : typeof err === 'string'
-              ? err
-              : 'unknown error';
-        pendingPlatformErrors.push({
-          tagname: 'cond',
-          message,
-          line: NaN,
-          column: NaN,
-          reason: message
-        });
-        return false;
-      }
-    },
-    'xstate.stateIn': (args: any, params: any) => {
-      const stateId = params?.stateId as string;
-      if (!stateId) return false;
-      const normalizedId = stateId.replace(/^#/, '');
-      const snapshot = args._snapshot;
-      if (snapshot?.nodes) {
-        return snapshot.nodes.some((node: any) => node.id === normalizedId);
-      }
-      return false;
-    },
-    'xstate.not': (args: any, params: any) => {
-      const inner = params?.guard;
-      const innerImpl = inner && args.guards?.[inner.type];
-      if (!innerImpl) {
-        throw new Error(
-          `Guard '${inner?.type}' referenced by 'xstate.not' is not implemented.`
-        );
-      }
-      return !innerImpl(args, inner.params);
-    }
-  };
   const provided = machine.provide({
     actions: resolvedSources.actions,
     actors: resolvedSources.actors,
     guards: {
-      ...providedGuards,
+      'xstate.stateIn': (args: any, params: any) => {
+        const stateId = params?.stateId as string;
+        const snapshot = args._snapshot;
+        return (
+          !!stateId &&
+          !!snapshot?.nodes?.some(
+            (node: any) => node.id === stateId.replace(/^#/, '')
+          )
+        );
+      },
+      'xstate.not': (args: any, params: any) => {
+        const inner = params?.guard;
+        const impl = inner && args.guards?.[inner.type];
+        if (!impl)
+          throw new Error(
+            `Guard '${inner?.type}' referenced by 'xstate.not' is not implemented.`
+          );
+        return !impl(args, inner.params);
+      },
       ...resolvedSources.guards
     },
     delays: resolvedSources.delays
   });
-
-  // Keep the original JSON so `serializeMachine(machine)`
-  // round-trip losslessly.
   (provided as any)._json = json;
   return provided;
 }
