+++ packages/core/src/scxml/index.ts
@@ -1,0 +1,4 @@
+export {
+  createMachineFromSCXML,
+  type SCXMLConversionOptions
+} from './scxml.ts';
